+++ src/ts/utils/helpers/VEllipseHelper.ts
@@ -1,9 +1,10 @@
 /**
  * A helper for VEllipses.
  *
- * @author  Ikaros Kappler
- * @date    2025-03-31
- * @version 1.0.0
+ * @author   Ikaros Kappler
+ * @date     2025-03-31
+ * @modified 2025-04-02 Adding `VEllipseHelper.drawHandleLines()`.
+ * @version  1.0.0
  */
 
 import { Line } from "../../Line";
@@ -40,8 +41,17 @@ export class VEllipseHelper {
     });
   }
 
+  /**
+   * Draw grey handle lines.
+   *
+   * @param {DrawLib<any>} draw - The draw library instance to use.
+   * @param {DrawLib<any>} fill - The fill library instance to use.
+   */
   drawHandleLines(draw: DrawLib<any>, fill: DrawLib<any>) {
     // Draw rotation handle line
-    draw.line(this.ellipse.center, this.rotationControlPoint, "rgba(64,192,128,0.333)", 1.0);
+    draw.line(this.ellipse.center, this.rotationControlPoint, "rgba(64,192,128,0.333)", 1.0, {
+      dashOffset: 0.0,
+      dashArray: [4, 2]
+    });
   }
 }